+++ lib/src/commands/gg_multi_do.dart
@@ -8,7 +8,6 @@ import 'package:args/command_runner.dart';
 import 'package:gg_log/gg_log.dart';
 import 'package:gg_multi/src/commands/do/add.dart';
 import 'package:gg_multi/src/commands/do/add_deps.dart';
-import 'package:gg_multi/src/commands/do/cancel_review.dart';
 import 'package:gg_multi/src/commands/do/checkout.dart';
 import 'package:gg_multi/src/commands/do/claude.dart';
 import 'package:gg_multi/src/commands/do/code.dart';
@@ -52,7 +51,6 @@ class Do extends Command<void> {
     addSubcommand(DoPublishCommand(ggLog: ggLog));
     addSubcommand(DoConfigurePublishCommand(ggLog: ggLog));
     addSubcommand(DoReviewCommand(ggLog: ggLog));
-    addSubcommand(DoCancelReviewCommand(ggLog: ggLog));
     addSubcommand(MaintainCommand(ggLog: ggLog));
     addSubcommand(DoInstallGitattributesCommand(ggLog: ggLog));
     addSubcommand(DoClaudeCommand(ggLog: ggLog));