+++ test/commands/gg_multi_do_test.dart
@@ -43,7 +43,6 @@ void main() {
           'push',
           'publish',
           'review',
-          'cancel-review',
         ],
       );